Move window creation into State::new()
